Combine the two sets of DeferredCache tests
